# Removed boxing and string allocations in MessageTemplateTextFormatter