@nblumhardt 2024-08-02
Looks like a great change, left a comment on the other version of this PR which I think will also apply here: https://github.com/serilog/serilog-expressions/pull/121

Cheers!